+++ screen2.0/src/app/applets/argo/sequence/sequenceTable.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from "react";
 import { MotifQueryDataOccurrence, SequenceTableProps, SequenceTableRow } from "../types";
 import { DataTable, DataTableColumn } from "@weng-lab/psychscreen-ui-components";
-import MotifsModal from "../motifModal";
+import MotifsModal from "./motifModal";
 
 const SequenceTable: React.FC<SequenceTableProps> = ({
     sequenceFilterVariables,
